Improve error handling and note ownership in noteController Revised error handling to provide more informative messages and implemented ownership verification for note CRUD operations to ensure secure access control. These changes enhance the user experience and security of the application. (Issue #123)

+++ backend/controllers/noteController.js
@@ -1,10 +1,14 @@
 import mongoose from "mongoose";
 import Note from "../models/noteModel.js";
 
+const error_msg =
+  "Apologies, the requested note could not be found. Please ensure the provided ID is accurate or consider exploring other available notes.";
+
 // GET all note
 const getNotes = async (req, res) => {
   try {
-    const notes = await Note.find({}).sort({ createdAt: -1 });
+    const createdBy = req.user._id;
+    const notes = await Note.find({ createdBy }).sort({ createdAt: -1 });
     res.status(200).json(notes);
   } catch (error) {
     res.status(400).json({ error: error.message });
@@ -14,12 +18,13 @@ const getNotes = async (req, res) => {
 // GET a single note
 const getNote = async (req, res) => {
   try {
+    const createdBy = req.user._id;
     const { id } = req.params;
     if (!mongoose.Types.ObjectId.isValid(id))
-      return res.status(404).json({ error: "No such note" });
+      return res.status(404).json({ error: error_msg });
 
-    const note = await Note.findById(id);
-    if (!note) return res.status(400).json({ errror: "No such note" });
+    const note = await Note.findOne({ _id: id, createdBy });
+    if (!note) return res.status(400).json({ errror: error_msg });
     res.status(200).json(note);
   } catch (error) {
     res.status(400).json({ error: error.message });
@@ -29,7 +34,10 @@ const getNote = async (req, res) => {
 // CREATE a new note
 const createNote = async (req, res) => {
   try {
-    const note = await Note.create(req.body);
+    const createdBy = req.user._id;
+
+    const note = await Note.create({ ...req.body, createdBy });
+
     res.status(200).json(note);
   } catch (error) {
     res.status(400).json({ error: error.message });
@@ -39,11 +47,12 @@ const createNote = async (req, res) => {
 // DELETE a note
 const deleteNote = async (req, res) => {
   try {
+    const createdBy = req.user._id;
     const { id } = req.params;
     if (!mongoose.Types.ObjectId.isValid(id))
-      return res.status(400).json({ errror: "No such note" });
-    const note = await Note.findOneAndDelete({ _id: id });
-    if (!note) return res.status(400).json({ errror: "No such note" });
+      return res.status(400).json({ errror: error_msg });
+    const note = await Note.findOneAndDelete({ _id: id, createdBy });
+    if (!note) return res.status(400).json({ errror: error_msg });
     res.status(200).json(note);
   } catch (error) {
     res.status(400).json({ error: error.message });
@@ -55,9 +64,9 @@ const updateNote = async (req, res) => {
   try {
     const { id } = req.params;
     if (!mongoose.Types.ObjectId.isValid(id))
-      return res.status(400).json({ errror: "No such note" });
+      return res.status(400).json({ errror: error_msg });
     const note = await Note.findOneAndUpdate({ _id: id }, { ...req.body });
-    if (!note) return res.status(400).json({ errror: "No such note" });
+    if (!note) return res.status(400).json({ errror: error_msg });
     res.status(200).json(note);
   } catch (error) {
     res.status(400).json({ error: error.message });
